rework event processing - added delete event table liquibase script - Event is no longer an entity - removed NoArgsConstructor in Event - replaced SuperBuilder with Builder in Event - removed callSuper param from EqualsAndHashCode in Event - removed EventRepository - moved Event to events package - changed event variable type to var in publishEvent method in EventPublisher - events are no longer saved to local database in EventProcessor - removed unused log field in SpaceEntityService - changed type of enabled field to primitive boolean in SchedulingProperties - enabled field now has a default value of true in SchedulingProperties - removed validation from enabled field in SchedulingProperties - fixed spaceEntityStatisticsJobCron field javadoc in in SchedulingProperties - removed EventProcessorTest - moved some abstract test classes to abstractions package - changed springdoc section position in bootstrap.properties - changed spring-boot-starter-amqp dependency scope to implementation in build.gradle - added lombok compileOnly dependency in build.gradle
