+++ src/main/java/com/example/universe/simulator/entityservice/events/Event.java
@@ -1,26 +1,23 @@
-package com.example.universe.simulator.entityservice.entities;
+package com.example.universe.simulator.entityservice.events;
 
 import com.example.universe.simulator.entityservice.types.EventType;
+import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import lombok.experimental.SuperBuilder;
 
 import javax.persistence.Column;
-import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
+import java.io.Serializable;
 import java.time.OffsetDateTime;
 
-@Entity
 @Getter @Setter
-@NoArgsConstructor
-@SuperBuilder
-@EqualsAndHashCode(callSuper = true)
+@Builder
+@EqualsAndHashCode
 @ToString
-public class Event extends AbstractEntity {
+public class Event implements Serializable {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
