+++ src/main/java/com/example/universe/simulator/entityservice/events/EventProcessor.java
@@ -1,7 +1,5 @@
 package com.example.universe.simulator.entityservice.events;
 
-import com.example.universe.simulator.entityservice.entities.Event;
-import com.example.universe.simulator.entityservice.repositories.EventRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
@@ -13,12 +11,9 @@ import org.springframework.stereotype.Component;
 @Slf4j
 public class EventProcessor {
 
-    private final EventRepository eventRepository;
-
     @EventListener
     @Async
     public void process(Event event) {
-        eventRepository.save(event);
         log.info("processed {}", event);
     }
 }
