+++ src/main/java/com/example/universe/simulator/entityservice/events/EventPublisher.java
@@ -1,6 +1,5 @@
 package com.example.universe.simulator.entityservice.events;
 
-import com.example.universe.simulator.entityservice.entities.Event;
 import com.example.universe.simulator.entityservice.types.EventType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +16,7 @@ public class EventPublisher {
     private final ApplicationEventPublisher applicationEventPublisher;
 
     public void publishEvent(EventType type, String data) {
-        Event event = Event.builder()
+        var event = Event.builder()
             .type(type)
             .data(data)
             .time(OffsetDateTime.now())
