--- src/main/java/com/example/universe/simulator/entityservice/repositories/EventRepository.java
@@ -1,8 +1,0 @@
-package com.example.universe.simulator.entityservice.repositories;
-
-import com.example.universe.simulator.entityservice.entities.Event;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface EventRepository extends JpaRepository<Event, UUID> {}
