+++ src/main/resources/bootstrap.properties
@@ -30,6 +30,9 @@ spring.datasource.password=${postgres-password}
 #jpa
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.open-in-view=false
+#springdoc
+springdoc.swagger-ui.tags-sorter=alpha
+springdoc.swagger-ui.operations-sorter=alpha
 #rabbitmq
 spring.rabbitmq.host=${rabbitmq-host}
 spring.rabbitmq.port=${rabbitmq-port}
@@ -38,8 +41,5 @@ spring.rabbitmq.password=${rabbitmq-password}
 spring.rabbitmq.ssl.enabled=true
 spring.rabbitmq.ssl.trust-store=${server.ssl.key-store}
 spring.rabbitmq.ssl.trust-store-password=${US_KEYSTORE_PASSWORD}
-#springdoc
-springdoc.swagger-ui.tags-sorter=alpha
-springdoc.swagger-ui.operations-sorter=alpha
 #zipkin
 spring.zipkin.base-url=http://${zipkin-host}:${zipkin-port}
