+++ src/test/java/com/example/universe/simulator/entityservice/common/abstractions/AbstractTest.java
@@ -1,4 +1,4 @@
-package com.example.universe.simulator.entityservice.common;
+package com.example.universe.simulator.entityservice.common.abstractions;
 
 import org.springframework.test.context.ActiveProfiles;
