+++ src/test/java/com/example/universe/simulator/entityservice/integration/AbstractIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.example.universe.simulator.entityservice.integration;
 
-import com.example.universe.simulator.entityservice.common.AbstractMockMvcTest;
-import com.example.universe.simulator.entityservice.common.AbstractSpringBootTest;
+import com.example.universe.simulator.entityservice.common.abstractions.AbstractMockMvcTest;
+import com.example.universe.simulator.entityservice.common.abstractions.AbstractSpringBootTest;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 
 @AbstractSpringBootTest
