+++ src/test/java/com/example/universe/simulator/entityservice/unit/AbstractWebMvcTest.java
@@ -1,7 +1,7 @@
 package com.example.universe.simulator.entityservice.unit;
 
-import com.example.universe.simulator.entityservice.common.AbstractMockMvcTest;
-import com.example.universe.simulator.entityservice.common.AbstractTest;
+import com.example.universe.simulator.entityservice.common.abstractions.AbstractMockMvcTest;
+import com.example.universe.simulator.entityservice.common.abstractions.AbstractTest;
 import com.example.universe.simulator.entityservice.config.BeanConfig;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.modelmapper.ModelMapper;
